pyproject.toml
# =====================================================
# Project Settings
# =====================================================
[project]
name = "capy-discord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = { text = "MIT" }
requires-python = "==3.13.*"
dependencies = ["discord-py>=2.6.4", "pydantic-settings>=2.12.0"]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "ruff>=0.14.5",
    "ty>=0.0.1a26",
    "uv>=0.9.10",
    "taskipy>=1.14.1",
    "pytest>=9.0.1",
    "pytest-xdist>=3.8.0",
    "pytest-cov>=7.0.0",
    "coverage>=7.12.0",
]

# =====================================================
# Scripts
# =====================================================
[tool.taskipy.tasks]
start = { cmd = "python -m capy_discord", help = "Runs the python application locally." }
lint = { cmd = "pre-commit install && pre-commit run --all-files", help = "Installs git hooks if needed, then lints and formats the codebase." }
build = { cmd = "docker build -t capy-discord .", help = "Builds the application's Docker image." }
run = { cmd = "docker run -it --rm capy-discord", help = "Runs the application inside a new Docker container." }
dev = { cmd = "docker run -it --rm -v ./bot:/app/bot capy-discord", help = "Runs the app in Docker with local code mounted for live changes." }
test = { cmd = "pytest -n auto --ff", help = "Runs all tests, starting with previously failed ones." }
retest = { cmd = "pytest -n auto --lf", help = "Reruns only the tests that failed during the last run." }
testcov = { cmd = "pytest -n auto --cov=bot --cov-report= && coverage report", help = "Runs tests, generates coverage data, and fails if coverage is below 80%." }
commit = { cmd = "pre-commit install && git add . && git commit -m", help = "Stages all changes and commits with message." }

# =====================================================
# Ruff Linting Configuration
# =====================================================
[tool.ruff]
target-version = "py313"
line-length = 120
output-format = "concise"
unsafe-fixes = true

[tool.ruff.lint]
select = [
    # --- Core & Syntax (The Basics) ---
    "F",   # Pyflakes: Catches syntax errors and undefined names
    "E",   # pycodestyle Error: Logical and style errors
    "W",   # pycodestyle Warning: Style warnings
    "C90", # McCabe: Flags functions that are too complex (cyclomatic complexity)

    # --- Modernization & Formatting ---
    "I",  # isort: Sorts imports automatically
    "UP", # pyupgrade: Enforces modern Python syntax (e.g., f-strings over .format())
    "D",  # pydocstyle: Enforces docstring conventions (Google/NumPy style)

    # --- Best Practices (Bug Prevention) ---
    "B",   # flake8-bugbear: Catches common bugs and design flaws
    "SIM", # flake8-simplify: Suggests cleaner, simpler code patterns
    "C4",  # flake8-comprehensions: Enforces better list/dict/set comprehensions
    "A",   # flake8-builtins: Prevents shadowing built-ins (like naming a var 'list')
    "RET", # flake8-return: Enforces consistent return statements

    # --- Strict Typing & Annotations ---
    "ANN", # flake8-annotations: FORCES type annotations on everything
    "TCH", # flake8-type-checking: Moves typing-only imports behind 'if TYPE_CHECKING:'
    "N",   # pep8-naming: Enforces strict variable/class naming conventions
    "ARG", # flake8-unused-arguments: flags unused function arguments

    # --- Safety, Security & Reliability ---
    "S",   # flake8-bandit: Security testing (e.g., hardcoded passwords)
    "DTZ", # flake8-datetimez: Bans naive datetimes (must use timezones)
    "TRY", # tryceratops: Enforces better exception handling practices

    # --- Performance & Logging ---
    "PERF", # Perflint: Checks for performance anti-patterns
    "LOG",  # flake8-logging: Enforces logging best practices
    "T20",  # flake8-print: Bans 'print' statements (force use of logger)

    # --- Cleanliness & Refactoring ---
    "ERA", # eradicate: Bans commented-out code
    "PL",  # Pylint: The "Strict" refactoring rules (PLR), conventions (PLC), etc.
    "RUF", # Ruff-specific: Custom rules from the Ruff team (often new best practices)

    # --- Testing ---
    "PT", # flake8-pytest-style: Enforces best practices for pytest
]
ignore = [
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "D", "S101"]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
order-by-type = false
case-sensitive = true
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 10
